refactor: merge filters folder into utils
